+++ src/converters/piral-vue-3/src/converter.ts
@@ -2,6 +2,7 @@ import type { ForeignComponent, BaseComponentProps } from 'piral-core';
 import { Component, App } from 'vue';
 import { createExtension } from './extension';
 import { mountVue } from './mount';
+import { Vue3MiddlewareHandler } from './types';
 
 export interface Vue3ConverterOptions {
   /**
@@ -23,6 +24,7 @@ interface Vue3State {
 export function createConverter(config: Vue3ConverterOptions = {}) {
   const { rootName = 'piral-slot', selector = 'extension-component' } = config;
   const Extension = createExtension(rootName);
+  const middlewares: Array<Vue3MiddlewareHandler> = [];
   const convert = <TProps extends BaseComponentProps>(
     root: Component<TProps>,
     captured?: Record<string, any>,
@@ -30,8 +32,9 @@ export function createConverter(config: Vue3ConverterOptions = {}) {
     mount(parent, data, ctx, locals: Vue3State) {
       const el = parent.appendChild(document.createElement(rootName));
       const app = mountVue(root, data, ctx, captured);
-      app.mount(el);
+      middlewares.forEach((middleware) => middleware(app));
       app.component(selector, createExtension(rootName));
+      app.mount(el);
       !app._props && (app._props = {});
       locals.instance = app;
     },
@@ -47,5 +50,10 @@ export function createConverter(config: Vue3ConverterOptions = {}) {
     },
   });
   convert.Extension = Extension;
+  convert.defineMiddleware = (middleware: Vue3MiddlewareHandler) => {
+    if (!middlewares.includes(middleware)) {
+      middlewares.push(middleware);
+    }
+  };
   return convert;
 }